slm/requirements.txt
torch>=2.1.0
transformers>=4.40.0
trl>=0.8.0
peft>=0.10.0
unsloth>=2024.3
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
huggingface_hub>=0.22.0
gradio>=4.0.0
pyyaml>=6.0
pydantic>=2.0.0
